requirements.txt
python-telegram-bot>=20.0
requests>=2.28
python-dotenv>=1.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
